ci: uplift pull request checks
Removes the existing verify code, verify commits and verify pr metadata workflows in favor of a single check pr workflow. This workflow will run on the main branch or any pr targeting it, and check the following
- the code builds/compiles
- the code meets the eslint requirements
- all the unit tests pass and have an overall coverage of at least 95%
- the vuepress docs build
- the commits meets the commitlint requirements 
- the source code and documentation are spelled correctly

This is required to support the new workflow of re basing pull requests directly on to the main branch, then publishing a new release. As such these pull request checks need to be more stringent before allowing the pull request to be merged.