ci(publish-release): run yarn build before publishing a new release
Runs the `yarn build` command before running `yarn release`. Subsequently the `@semantic-release/exec` plugin is no longer required.

Previously the `yarn build` would run as a part of prepare step of semantic release using the `@semantic-release/exec` plugin. The issue with this is that the `@semantic-release/npm` plugin checks that the `dist/package.json` file exists before the command has run, causing semantic release to error out. By running the command before semantic release runs, the `dist/package.json` is guaranteed to exist.

The drawback to this is that the `yarn build` command will always run even if a new release isn't required.