chore: set distributed package.json to public
Updates the roll up config so that the generated package.json used for distribution, has private set to false. Meanwhile sets private to true in the package.json in the root of this repository, as this is used for development. This ensures that the built, minified code can only ever be published to NPM, but the source code cannot.