chore: copy version number from dist/package.json to package.json
Adds a copy-version.sh script which copies the version from the dist/package.json to the package.json in the root of the repo. This script is automatically ran by semantic release when it updates the version number in dist/package.json and is included in the commit automatically generated.